bigfin: Changed the disk size and journal sizes as float64
Bug-URL: https://bugzilla.redhat.com/1358627

+++ backend/backend.go
@@ -271,5 +271,5 @@ type DeviceDetail struct {
 	Uuid     uuid.UUID `json:"uuid"`
 	PartName string    `json:"partname"`
 	FSType   string    `json:"fstype"`
-	Size     uint64    `json:"size"`
+	Size     float64   `json:"size"`
 }
